// src/components/location-autocomplete/types.ts
export type FieldsT = {
  building_number?: string;
  level_number?: string;
  unit_type?: string;
  unit_number?: string;
  lot_number?: string;
  street_number?: string;
  street_name?: string;
  street_type?: string;
  suffix?: string;
  suburb?: string;
  state?: string;
  pincode?: string;
  lga?: string;
};